Separate FkAyat class
