+++ src/tests/unit/srv/ayats/test_neighbor_ayats.py
@@ -25,8 +25,9 @@ from typing import override
 import pytest
 
 from app_types.listable import AsyncListable
-from srv.ayats.ayat import Ayat, FkAyat
+from srv.ayats.ayat import Ayat
 from srv.ayats.ayat_identifier import FkIdentifier
+from srv.ayats.fk_ayat import FkAyat
 from srv.ayats.neighbor_ayats import FavoriteNeighborAyats
 from srv.files.file import FkFile
